[FEAT] 프로필 편집 함수 통합
EditProfileIcon과 ProfileEditText를 합쳐 EditableProfile을 생성함.

Resolves: TICO-66
Related to: TICO-61
