+++ Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/ui/auth/view/SignUpScreen.kt
@@ -6,23 +6,13 @@ import android.net.Uri
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.PickVisualMediaRequest
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.BasicTextField
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
@@ -33,13 +23,8 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.SolidColor
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
 import androidx.core.content.ContextCompat
 import androidx.core.content.FileProvider
@@ -49,11 +34,10 @@ import com.tico.pomorodo.BuildConfig
 import com.tico.pomorodo.R
 import com.tico.pomorodo.ui.auth.viewModel.AuthViewModel
 import com.tico.pomorodo.ui.common.view.CustomTextButton
-import com.tico.pomorodo.ui.common.view.EditProfileIcon
+import com.tico.pomorodo.ui.common.view.EditableProfile
 import com.tico.pomorodo.ui.common.view.createImageFile
 import com.tico.pomorodo.ui.common.view.executeToast
 import com.tico.pomorodo.ui.theme.PomoroDoTheme
-import com.tico.pomorodo.ui.theme.laundryGothic
 import java.util.Objects
 
 
@@ -176,12 +160,12 @@ fun SignUpScreen(
             style = PomoroDoTheme.typography.laundryGothicBold20
         )
         Spacer(modifier = Modifier.height(20.dp))
-        EditProfileIcon(
+        EditableProfile(
             profileUri = profileUri,
-            onProfileClicked = onProfileClicked
+            onProfileClicked = onProfileClicked,
+            inputText = inputText,
+            onInputTextChanged = onInputTextChanged
         )
-        Spacer(modifier = Modifier.height(18.dp))
-        ProfileEditText(inputText = inputText, onInputTextChanged = onInputTextChanged)
         Spacer(modifier = Modifier.height(20.dp))
         CustomTextButton(
             text = stringResource(id = R.string.content_button_sign_up),
@@ -193,72 +177,6 @@ fun SignUpScreen(
             textStyle = PomoroDoTheme.typography.laundryGothicRegular18,
             verticalPadding = 12.dp,
             onClick = onSignUpButtonClicked
-        )
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ProfileEditText(inputText: String, onInputTextChanged: (String) -> Unit) {
-    val focusManager = LocalFocusManager.current
-    val interactionSource = remember { MutableInteractionSource() }
-    val colors = TextFieldDefaults.colors(
-        focusedContainerColor = PomoroDoTheme.colorScheme.background,
-        unfocusedContainerColor = PomoroDoTheme.colorScheme.background,
-        focusedIndicatorColor = PomoroDoTheme.colorScheme.primaryContainer,
-        focusedLabelColor = PomoroDoTheme.colorScheme.primaryContainer,
-        focusedTextColor = PomoroDoTheme.colorScheme.onBackground,
-        unfocusedTextColor = PomoroDoTheme.colorScheme.onBackground,
-        cursorColor = PomoroDoTheme.colorScheme.primaryContainer
-    )
-
-    BasicTextField(
-        value = inputText,
-        onValueChange = { onInputTextChanged(it) },
-        modifier = Modifier
-            .fillMaxWidth()
-            .heightIn(min = 0.dp),
-        textStyle = PomoroDoTheme.typography.laundryGothicRegular16.copy(color = colors.unfocusedTextColor),
-        keyboardOptions = KeyboardOptions(
-            keyboardType = KeyboardType.Text,
-            imeAction = ImeAction.Done
-        ),
-        keyboardActions = KeyboardActions(onDone = { focusManager.clearFocus() }),
-        interactionSource = interactionSource,
-        cursorBrush = SolidColor(colors.cursorColor)
-    ) { innerTextField ->
-        TextFieldDefaults.DecorationBox(
-            value = inputText,
-            innerTextField = innerTextField,
-            enabled = true,
-            singleLine = true,
-            visualTransformation = VisualTransformation.None,
-            interactionSource = interactionSource,
-            isError = false,
-            label = {
-                Text(
-                    text = stringResource(R.string.content_user_name_label),
-                    fontFamily = laundryGothic
-                )
-            },
-            placeholder = {
-                Text(
-                    text = stringResource(R.string.content_user_name_placeholder),
-                    color = PomoroDoTheme.colorScheme.gray50,
-                    style = PomoroDoTheme.typography.laundryGothicRegular16
-                )
-            },
-            shape = RoundedCornerShape(size = 5.dp),
-            colors = colors,
-            container = {
-                TextFieldDefaults.Container(
-                    enabled = true,
-                    isError = false,
-                    interactionSource = interactionSource,
-                    colors = colors,
-                )
-            },
-            contentPadding = PaddingValues(horizontal = 12.dp, vertical = 14.dp)
         )
     }
 }
